+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.AppendFormat.cs
@@ -15,7 +15,6 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AppendFormat<T>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] scoped ReadOnlySpan<char> format,
         T arg)
@@ -71,7 +70,6 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AppendFormat<T1, T2>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] scoped ReadOnlySpan<char> format,
         T1 arg1,
@@ -139,7 +137,6 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AppendFormat<T1, T2, T3>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] scoped ReadOnlySpan<char> format,
         T1 arg1,
@@ -213,7 +210,6 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AppendFormat<T1, T2, T3, T4>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] scoped ReadOnlySpan<char> format,
         T1 arg1,
@@ -293,7 +289,6 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AppendFormat<T1, T2, T3, T4, T5>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] scoped ReadOnlySpan<char> format,
         T1 arg1,